updated Readme.md, and minor change in test
created a brief structure of the document readme.md, added short comments or memo for future creation of the sub sections
added a few pintouts in CashPlan() tests. the reason why I added these tests are because that the CashPlan() object can NOT be printed in the shell linked with Pyharm. problem NOT found and not solved