improved built-in strategy creation process
removed all "switch" or "if-elif-else" codes in strategy creation in operater.py. instead, a dict is created in built-in.py that contains all available strategies with a code name. and code names can be refered and coresponding strategies are added to operator object when strategy code is refered
changed code for some built-in strategies