updated operator.py to keep redundent signal
removing redundent signal was previous design. but this might leads to good signals being deleted: for example, a series of signal: 0.3, 0.3, 0.3 should be considered good signals because this is a step-by-step position build up, but this type of signals will be removed and only the first signal remains. original idea is to remove all the redundent signals to reduce number of signals and reduces time. but the time can not be saved unless proper way is found