corrected a mistake
a mistake has been detected and solved.
the mistake has lead to abnormal investment back testing signal start date, where the first operation signal has always been too far away from expected test start date. for example, in a back test with invest_start date is 2006-06-04, the first operation signal has always been later than 2007-jan, if the investment start date moves on wards, the first op signal moves on the same direction
the date bounds generated in selecting strategy is found to be the root cause of the problem. multiple "nan"signals are generated because the time delta between the actual time bounds and expected time bounds.
the problem is solved by offsetting the time bounds so that it overlaps with the expected one
a new test case TestFastExperiment is created to find out the root cause, and to serve as quick experiment tests