added asset_type and freq support in database.py
added asset)type and freq support in database.py
now different file names will be used for the files that store historical data of asset_type other than 'E' and historical data freq other than 'D'
for example, close prices of Indices like '000300.SH' the “沪深300指数" will be stored with file name close-I.csv. this these asset types and other freq s can be supported in database.py.
above functions are tested when file is created, there is still bug when the file is created for the first time. to be further tested