solved a problem in database
solved the problem of got wrong historical data from local when the file is being created for the first time, and then cause wrong result in optimization and back looping.

the problem was caused by the function get_and_update_date() in database.py, when a historical data frame is created for the first time, an empty DF is created, and this DF might not have exactly the same date range as the one to be extracted, and this DF is returned as the result. if its date range is different, then mistake will be caused.

the problem was solved by re-extract the historical data from local file while it is created, thus correct DF will anyway be returned, thus problem solved