corrected mistake in history.py, CONFUSED!!
corrected a very strange mistake:
if .sort_index() is removed from the code, it will fail, the df that has been regulated will be re-assigned to the regulate_df() function, and then program will fail because the df has been regulated already. error will be thrown when two already removed columns are tried to be removed again.
DON'T know why this is happenning, seems like if .sort_index() is put back in the code, there will be no problem.