improved run_mode2 in core.py, realized multiple period testing
realized multiple period testing by following modifications, and tested successfully:
extended function _get_parameter_performance() to cover both optimize and test stages of optimization, and unified evaluation processes, so that both stages can call the same function to get the results as performance of the parameter
the output of above function is then adjusted to be a dict instead of a float, the float performance shall be accessed through dict after performance calculation.
in evaluate.py, a function that calculates statistics of performances is added then can be used in multiple cases, one of the case is to calculate the average performance in multiple sub-period optimize or testing. later it can also be used in visualization or printing of results of optimization / testing. further combination of codes can be done