improved loop final value evaluation accuracy
1, improved the loop result evaluation output, ensured full range data being used in the backtest result, thus made the output result more accurate.
explanation:
previously the loop final value is not the total value at the last day of loop range, instead, the total value at the last day with operation signal.  for example, if the loop range starts from 2019/01/01 and ends at 2021/03/01, ideally the final value should be the value on 2021/03/01. however, it was not the case, since back test module calculates only the total values on the dates with operation signals, thus the standard loop result output was only on these operation dates, for example, the last operation date might be 2021/02/05, close to but not the last day of the loop range. clearly the final value on 2021/02/05 might NOT be the same as that on 2021/03/01. thus the "final value" on last operation date is only a "pseudo final value" which might different from the real one.
previously the pseudo final values are used to evaluate a strategy and its parameters because the calculation work is lower.
Now, the real final value can be calculated since more accurate evaluations should be done on the looped value, it does not make much sense to use the pseudo final value
by calling _get_complete_hist() function in _evaluate_one_parameter() we create the full range looped values, and thus made accurate result of evaluation always available.
2, improved printings of the back test result:
2.1 removed extra "ms" for run timing
2.2 limited the number of digits for looped periods