partially solved problem of opti and test start time wrong
the problem was caused by wrong starting date setting for optimization and testing period test-looping.
instead of passing first cash dates as the start point, the first day of loop historical data is used. thus empty signal dates are included in the loop results
now correct date is used, however the problem is only partially solved, because when the cash_dates are given as list, current code will not work