corrected a bug that causes Nan data being used in loop
a mistake in database.py is detected that might cause NaN values being passed to operation signal generation and looping. when np.inf existed in the data, and they can not be properly removed as nan values. the correction is simple: to tell pandas to treat inf values as nan (temporarily) by setting the option "use_inf_as_na" to be True, see row 267 and 268 in database.py

this will solve the problem of downloading data contains inf and then cause problems

other changes are minor, including:
1, removed try...except... in multiprocessing of calling pro_bar in history.py
2, added first try of logging module