modified test to find out how large data can be downloaded by multiprocess
it shows that if each downloaded data block is too large, the concurrent future module will throw queue.FULL error, and it seem that this error is related to the size of data block each process is working on, and there's no solution to it.
it shows that if too much data is downloaded, when the time period exceeds about 27 years, it is likely to have queue.FULL error. now testing with smaller set of data downloaded