started to modify class structure of Operator()
the Operator class is the main realization of strategies that utilizes them to create operation signals. in the old design the Operator class is designed so that different usages of strategies are adopted and all signals are merged into one. however, this structure has many drawbacks, one of which is the inflexibility to cope with different types of operation signals such as target signals or buysell signals. and multiple operation prices are not supported as well
now the improvement and changing of structure of Operator class will allow more flexibility of the signals
the current changes include:
1, modified introduction docstring in the Operator Class
2, the basic data structure of Operator which is defined in __init__() function
3, refactor init variables of the Operator
the modification has just started and will be tested when the modification is more or less completed