introduced multi-signal test in TestLoop to test multiple signal
multiple signal is the last new feature to be added to the new loop structure, that allows multiple signals to be looped in the same operation cycle at different prices. the key here is to ensure date-awareness of stock and cash delivery.
a manual test case is built up and new test case will be created to verify the features related to multiple signals.
after this the signal looping and signal creation (in operator class) will be fully connected in new structure and the QT module can be jointly tested for full functionality, and after that, the upgrading of qt to the new structure will be completed finally!!