added sample index check in signal creation in operator
prevents signal from being created when a strategy is not supposed to be activated, or nor sampled at the time
a strategy will only run when the hdate idx is in the strategy's sample index list
before this improvement, the strategy is always run in step mode, this is not correct, it should not always run. this behavior is to be further tested.