improved definition of strategy
added a new property: use_latest_data_cycle: bool, default True for strategy class:
controls if latest data cycle can be used to create trade signals: if not, only data from previous cycle can be used. if strategy is running at very low frequency, this will lead to trading delay.
however, in live trade, latest data cycle is not always available, like the close price of a daily candle is only available after the market is closed, thus trading is meaningless if the trader waits for the close price to be available. in this case, the latest close price will be estimated with latest realtime price, if the strategy is running at close timing (actually a few minutes before closing), the estimation will not be very wrong.
in backtest mode, the close price of current cycle of cause is always available, but the user still need to be aware of the danger of creating future function, and try cautiously to avoid such function