fixed a bug in fetch_real_time_klines
in the original code, there's a tiny bit of risk that the realtime k-line data extracted are with different time-stamps, that is, prices for different symbols are of different time, this might lead to mis-alignment of price data that are passed into strategy module and cause NaN value in the output.
Now, if the "matured_kline_only" is set to True, then the price time stamp will be compared with current time before return, thus make sure that the same price timestamp is returned from the time (unless...there are missing timestamps in the index) instead of passing the second last timestemp.

this will solve largely the problem

An example of the problem is as following:
if realtime k line for symbol A and B are extracted at 13:38 sharp, the original code will return the last row of each symbol in normal case and the second last row in matured only case:
if we extracted df for symbol A
13:35 45.0 - matured
13:40 45.5 - not mature
and for symbol B:
13:35 5.5
13:40 5.6
then return the last one in normal case and the second last one in MKO case is correct
but occassionally, i.e. at sharp 13:40, following data will be extracted:
for symbol A
13:35 45
13:40 45.5
but for symbol B
13:35:5.5
13:40 5.6
13:45 5.6
Note that there's one more row be extracted for symbol B but not for symbol A
in this case if we return the last row in normal case and the second last row in MKO case will result symbol A and symbol B having different time stamp!!!
To solve this problem the current time is compared with the df list and in MKO case only the row before current time is returned, thus A and B will have exactly same timestamp in MKO case !!! (Note, assuming that A and B does not have any time stamp missing in their indexes)
and in normal case it is allowed that their time stamps are different!